Add project source code formatting

+++ src/main/java/io/tiledb/spark/datasourcev2/TileDBWriterFactory.java
@@ -1,6 +1,14 @@
 package io.tiledb.spark.datasourcev2;
 
+import static io.tiledb.java.api.Layout.TILEDB_UNORDERED;
+import static io.tiledb.java.api.QueryType.TILEDB_WRITE;
+
 import io.tiledb.java.api.*;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Optional;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SaveMode;
 import org.apache.spark.sql.sources.v2.DataSourceOptions;
@@ -11,15 +19,6 @@ import org.apache.spark.sql.sources.v2.writer.WriterCommitMessage;
 import org.apache.spark.sql.types.StructType;
 import scala.collection.Seq;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Optional;
-
-import static io.tiledb.java.api.Layout.TILEDB_UNORDERED;
-import static io.tiledb.java.api.QueryType.TILEDB_WRITE;
-
 public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter {
   private transient Context ctx;
   private String jobId;
@@ -27,7 +26,12 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
   private boolean createTable, deleteTable;
   private TileDBOptions tileDBOptions;
 
-  public TileDBWriterFactory(String jobId, StructType schema, boolean createTable, boolean deleteTable, TileDBOptions tileDBOptions) {
+  public TileDBWriterFactory(
+      String jobId,
+      StructType schema,
+      boolean createTable,
+      boolean deleteTable,
+      TileDBOptions tileDBOptions) {
     this.jobId = jobId;
     this.schema = schema;
     this.createTable = createTable;
@@ -35,23 +39,29 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
     this.tileDBOptions = tileDBOptions;
   }
 
-  public TileDBWriterFactory(String jobId, StructType schema, boolean createTable, boolean deleteTable, TileDBOptions tileDBOptions, boolean init) throws Exception {
+  public TileDBWriterFactory(
+      String jobId,
+      StructType schema,
+      boolean createTable,
+      boolean deleteTable,
+      TileDBOptions tileDBOptions,
+      boolean init)
+      throws Exception {
     this.jobId = jobId;
     this.schema = schema;
     this.createTable = createTable;
     this.deleteTable = deleteTable;
     this.tileDBOptions = tileDBOptions;
     this.ctx = new Context();
-    if(init) {
-      if (deleteTable)
-        deleteTable();
+    if (init) {
+      if (deleteTable) deleteTable();
 
-      if (createTable)
-        createTable();
+      if (createTable) createTable();
     }
   }
 
-  public static Optional<DataSourceWriter> getWriter(String jobId, StructType schema, SaveMode mode, DataSourceOptions options) {
+  public static Optional<DataSourceWriter> getWriter(
+      String jobId, StructType schema, SaveMode mode, DataSourceOptions options) {
     boolean createTable = false, deleteTable = false;
     TileDBOptions tileDBOptions = null;
     Context ctx;
@@ -61,27 +71,32 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
       Array array = new Array(ctx, tileDBOptions.ARRAY_URI, TILEDB_WRITE);
       array.close();
     } catch (Exception tileDBError) {
-        createTable = true;
+      createTable = true;
     }
     try {
-      switch (mode){
+      switch (mode) {
         case Append:
-          return Optional.of(new TileDBWriterFactory(jobId, schema, createTable, false, tileDBOptions, true));
+          return Optional.of(
+              new TileDBWriterFactory(jobId, schema, createTable, false, tileDBOptions, true));
         case Overwrite:
-          if(!createTable)
-            return Optional.of(new TileDBWriterFactory(jobId, schema, true, true, tileDBOptions, true));
+          if (!createTable)
+            return Optional.of(
+                new TileDBWriterFactory(jobId, schema, true, true, tileDBOptions, true));
           else
-            return Optional.of(new TileDBWriterFactory(jobId, schema, true, false, tileDBOptions, true));
+            return Optional.of(
+                new TileDBWriterFactory(jobId, schema, true, false, tileDBOptions, true));
         case ErrorIfExists:
-          if(!createTable)
-            return Optional.empty();
+          if (!createTable) return Optional.empty();
           else
-            return Optional.of(new TileDBWriterFactory(jobId, schema, createTable, deleteTable, tileDBOptions, true));
+            return Optional.of(
+                new TileDBWriterFactory(
+                    jobId, schema, createTable, deleteTable, tileDBOptions, true));
         case Ignore:
-          if(!createTable)
-            return Optional.empty();
+          if (!createTable) return Optional.empty();
           else
-            return Optional.of(new TileDBWriterFactory(jobId, schema, createTable, deleteTable, tileDBOptions, true));
+            return Optional.of(
+                new TileDBWriterFactory(
+                    jobId, schema, createTable, deleteTable, tileDBOptions, true));
       }
     } catch (Exception tileDBError) {
       tileDBError.printStackTrace();
@@ -100,14 +115,12 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
   }
 
   @Override
-  public void commit(WriterCommitMessage[] messages) {
-  }
+  public void commit(WriterCommitMessage[] messages) {}
 
   @Override
-  public void abort(WriterCommitMessage[] messages) {
-  }
+  public void abort(WriterCommitMessage[] messages) {}
 
-  private void deleteTable() throws Exception{
+  private void deleteTable() throws Exception {
     TileDBObject.remove(ctx, tileDBOptions.ARRAY_URI);
   }
 
@@ -119,9 +132,9 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
   }
 
   class TileDBWriter implements DataWriter<Row> {
-    private HashMap<String,Pair<NativeArray,NativeArray>> nativeArrays;
+    private HashMap<String, Pair<NativeArray, NativeArray>> nativeArrays;
     private List<Row> batch;
-    private HashMap<String,Pair<Integer,Integer>> varLengthIndex;
+    private HashMap<String, Pair<Integer, Integer>> varLengthIndex;
     private int rowIndex;
     private TileDBOptions options;
     private StructType schema;
@@ -134,7 +147,7 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
     private ArrayList<Long> attributeCellValNum;
     private ArrayList<Datatype> attributeDatatype;
 
-    public TileDBWriter(StructType schema, TileDBOptions tileDBOptions)  {
+    public TileDBWriter(StructType schema, TileDBOptions tileDBOptions) {
       this.schema = schema;
       this.options = tileDBOptions;
       try {
@@ -178,20 +191,19 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
         for (int i = 0; i < attributeNames.size(); i++) {
           String name = attributeNames.get(i);
           long cellValNum = attributeCellValNum.get(i);
-          if(cellValNum != 1){ //array
+          if (cellValNum != 1) { // array
             try {
               Seq seq = (Seq) record.getAs(name);
               increaseRowIndex(name, seq.size());
-            } catch (ClassCastException e){
+            } catch (ClassCastException e) {
               byte[] seq = ((String) record.getAs(name)).getBytes();
               increaseRowIndex(name, seq.length);
             }
-          }
-          else{
-            increaseRowIndex(name,1);
+          } else {
+            increaseRowIndex(name, 1);
           }
         }
-        if(batch.size() >= options.BATCH_SIZE){
+        if (batch.size() >= options.BATCH_SIZE) {
           flush();
         }
       } catch (Exception tileDBError) {
@@ -201,29 +213,29 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
     }
 
     private void increaseValueIndex(String name, int length) {
-      Pair<Integer,Integer> index = getIndex(name);
+      Pair<Integer, Integer> index = getIndex(name);
       Integer second = index.getSecond();
-      second+=length;
+      second += length;
       index.setSecond(second);
     }
 
     private void increaseValueIndex(String name) {
-      Pair<Integer,Integer> index = getIndex(name);
+      Pair<Integer, Integer> index = getIndex(name);
       Integer second = index.getSecond();
       second++;
       index.setSecond(second);
     }
 
     private void increaseRowIndex(String name, int size) {
-      Pair<Integer,Integer> index = getIndex(name);
+      Pair<Integer, Integer> index = getIndex(name);
       Integer first = index.getFirst();
-      first+=size;
+      first += size;
       index.setFirst(first);
     }
 
-    private Pair<Integer,Integer> getIndex(String name) {
+    private Pair<Integer, Integer> getIndex(String name) {
       Pair<Integer, Integer> index = varLengthIndex.get(name);
-      if(index == null) {
+      if (index == null) {
         index = new Pair<>(0, 0);
         varLengthIndex.put(name, index);
       }
@@ -231,26 +243,27 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
     }
 
     private void flush() throws Exception {
-      if(batch.size() == 0)
-        return;
+      if (batch.size() == 0) return;
 
-//      System.out.println("Flushing: "+batch);
+      //      System.out.println("Flushing: "+batch);
 
       // Create query
       query = new Query(array, TILEDB_WRITE);
       query.setLayout(TILEDB_UNORDERED);
-      //NativeArray nsubarray = new NativeArray(ctx, subarrayBuilder.getSubArray(), arraySchema.getDomain().getType());
-      //query.setSubarray(nsubarray);
+      // NativeArray nsubarray = new NativeArray(ctx, subarrayBuilder.getSubArray(),
+      // arraySchema.getDomain().getType());
+      // query.setSubarray(nsubarray);
       nativeArrays = new HashMap<>();
       int ndim = dimensionNames.size();
       int nattr = attributeNames.size();
-      for (int i = 0; i < nattr; i ++) {
+      for (int i = 0; i < nattr; i++) {
         String attrName = attributeNames.get(i);
         Datatype attrType = attributeDatatype.get(i);
         long cellValNum = attributeCellValNum.get(i);
         if (cellValNum == Constants.TILEDB_VAR_NUM) {
           NativeArray first = new NativeArray(ctx, batch.size(), Datatype.TILEDB_UINT64);
-          NativeArray second = new NativeArray(ctx, (int) varLengthIndex.get(attrName).getFirst(), attrType);
+          NativeArray second =
+              new NativeArray(ctx, (int) varLengthIndex.get(attrName).getFirst(), attrType);
           Pair<NativeArray, NativeArray> pair = new Pair<>(first, second);
           nativeArrays.put(attrName, pair);
           query.setBuffer(attrName, first, second);
@@ -262,7 +275,7 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
         }
       }
       NativeArray coords;
-      try (Domain domain = arraySchema.getDomain())  {
+      try (Domain domain = arraySchema.getDomain()) {
         coords = new NativeArray(ctx, batch.size() * ndim, domain.getType());
       }
       nativeArrays.put(Constants.TILEDB_COORDS, new Pair<>(null, coords));
@@ -271,7 +284,7 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
       rowIndex = 0;
       varLengthIndex = new HashMap<>();
       for (Row record : batch) {
-        for(int dimIdx = 0 ; dimIdx < ndim; dimIdx++) {
+        for (int dimIdx = 0; dimIdx < ndim; dimIdx++) {
           String dimensionName = dimensionNames.get(dimIdx);
           coords.setItem(rowIndex * ndim + dimIdx, record.getAs(dimensionName));
         }
@@ -279,7 +292,7 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
           String attrName = attributeNames.get(attrIdx);
           long cellValNum = attributeCellValNum.get(attrIdx);
           Pair<NativeArray, NativeArray> pair = nativeArrays.get(attrName);
-          if(cellValNum == Constants.TILEDB_VAR_NUM) {
+          if (cellValNum == Constants.TILEDB_VAR_NUM) {
             int datatypeBytes = attributeDatatype.get(attrIdx).getNativeSize();
             try {
               Seq array = (Seq) record.getAs(attrName);
@@ -288,28 +301,27 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
                 increaseValueIndex(attrName);
               }
               pair.getFirst().setItem(rowIndex, (long) getIndex(attrName).getFirst());
-              increaseRowIndex(attrName,array.size() * datatypeBytes);
+              increaseRowIndex(attrName, array.size() * datatypeBytes);
             } catch (ClassCastException e) {
               String s = (String) record.getAs(attrName);
               int nbytes = s.getBytes().length;
               pair.getSecond().setItem(getIndex(attrName).getSecond(), s);
               increaseValueIndex(attrName, nbytes);
               pair.getFirst().setItem(rowIndex, (long) getIndex(attrName).getFirst());
-              increaseRowIndex(attrName,nbytes * datatypeBytes);
+              increaseRowIndex(attrName, nbytes * datatypeBytes);
             }
           } else {
             if (cellValNum == 1) {
               pair.getSecond().setItem(rowIndex, record.getAs(attrName));
-            }
-            else {
+            } else {
               try {
                 Seq array = (Seq) record.getAs(attrName);
                 for (int index = 0; index < cellValNum; index++) {
-                  pair.getSecond().setItem(rowIndex * (int)cellValNum + index, array.apply(index));
+                  pair.getSecond().setItem(rowIndex * (int) cellValNum + index, array.apply(index));
                 }
-              } catch (ClassCastException e){
+              } catch (ClassCastException e) {
                 String s = (String) record.getAs(attrName);
-                pair.getSecond().setItem(rowIndex * (int)cellValNum, s);
+                pair.getSecond().setItem(rowIndex * (int) cellValNum, s);
               }
             }
           }
@@ -337,9 +349,6 @@ public class TileDBWriterFactory implements DataWriterFactory, DataSourceWriter 
     }
 
     @Override
-    public void abort() throws IOException {
-    }
+    public void abort() throws IOException {}
   }
-
-
 }
